nicer conversion into the RAC world with a shorter utility function, Pods work
