+++ Cartfile.resolved
@@ -1,2 +1,2 @@
-github "czechboy0/BuildaUtils" "v0.0.10"
+github "czechboy0/BuildaUtils" "v0.0.11"
 github "czechboy0/DVR" "4e036d99d23aa3f8349c2648dda876112939d41a"
