--- XCSReactiveCocoaOSX/XCSReactiveCocoaOSX.h
@@ -1,19 +1,0 @@
-//
-//  XCSReactiveCocoaOSX.h
-//  XCSReactiveCocoaOSX
-//
-//  Created by Honza Dvorsky on 11/09/2015.
-//  Copyright © 2015 Honza Dvorsky. All rights reserved.
-//
-
-#import <Cocoa/Cocoa.h>
-
-//! Project version number for XCSReactiveCocoaOSX.
-FOUNDATION_EXPORT double XCSReactiveCocoaOSXVersionNumber;
-
-//! Project version string for XCSReactiveCocoaOSX.
-FOUNDATION_EXPORT const unsigned char XCSReactiveCocoaOSXVersionString[];
-
-// In this header, you should import all the public headers of your framework using statements like #import <XCSReactiveCocoaOSX/PublicHeader.h>
-
-
